pyproject.toml
[project]
name = "biobagel"
version = "0.1.12"
authors = [
  {name = "Jakub Lála", email = "jakublala@gmail.com"},
  {name = "Ayham Al-Saffar", email = "ayham.saffar@gmail.com"},
  {name = "Stefano Angioletti-Uberti", email = "sangiole@imperial.ac.uk"},
]
description = "Protein Engineering via Exploration of an Energy Landscape"
readme = "README.md"
license = "MIT"
license_files = "LICENSE"
keywords = ["protein", "protein design", "optimization", "monte carlo", "sampling", "energy landscape"]

requires-python = ">=3.11,<3.14"
dependencies = [
    "biotite>=1.0.1",
    "boileroom==0.2.2",
    "numpy>=2.2.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "matplotlib>=3.10.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
bagel = "bagel.cli:main"

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "mypy>=1.14.1",
    "pytest>=8.3.4",
    "pytest-xdist>=3.6.1",
    "pytest-cov>=6.0.0",
    "pytest-repeat>=0.9.4",
    "python-dotenv>=1.2.1",
    "ruff>=0.8.6",
    "sphinx>=8.1.3",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.1",
    "pre-commit>=3.6.0",
]
local = [
    "transformers>=4.49.0",
]
wandb = [
    "wandb>=0.23.0",
]

[project.urls]
Repository = "https://github.com/softnanolab/bagel.git"
Issues = "https://github.com/softnanolab/bagel/issues"

[tool.mypy]
strict = true
show_traceback = true
plugins = "numpy.typing.mypy_plugin"
files = ["**/src/bagel/*.py"] #selects files checked during mypy call, but does not affect the editor's MyPy linter
explicit_package_bases = true #allows giving directories different mypy settings

[[tool.mypy.overrides]]
module = ["biotite.*", "pytest.*", "pandas.*", "bagel.*", "dotenv", "wandb"] # libraries that don't have type stubs
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"] #tells editor's mypy linter to ignore these modules
ignore_errors = true

[tool.ruff]
line-length = 120
include = ["pyproject.toml", "**/src/bagel/*.py", "**/tests/*.py"]
fix = true

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = ["F403"] #allows wildcard import of pytest fixtures in test files

[tool.ruff.lint]
select = [
  "W", # automatically removes trailing whitespace, removes extra empty lines at end of files...
]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bagel"]

[tool.hatch.metadata]
allow-direct-references = true
